requirements.txt
fastapi
uvicorn[standard]
chromadb
sentence-transformers
onnxruntime
requests
python-multipart
pypdf
python-docx
numpy
supabase
python-dotenv
